refactor(webui): move routing-rule event-type mapping into lib
The shared mapping (`EVENT_TYPES`, `EventTypeOption`, `findEventType`,
`eventTypeLabel`) is non-component logic and shouldn't live in a
component file — moves it to a new `src/lib/routing-rules.ts`, matching
the existing pattern of `lib/task.ts`. The form keeps the
component-only `mentionCopyForSource` helper and imports the shared
pieces from `@/lib/routing-rules`. The Events page imports
`eventTypeLabel` from the same module instead of from the form.
